moved arpTable management out of endpoints.js to arpTable.js
